+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### 2020-11-11
+- Refactoring js.nameOf-methods
+
 ### 2020-11-05
 - Introducing fast-xml-parser (moved from cavalion-code)